backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.1
supabase==2.9.1
groq==0.11.0
sentence-transformers==3.3.1
flask==3.1.0
flask-cors==5.0.0
requests==2.32.3
pydantic==2.9.0
python-multipart==0.0.12
numpy==1.26.4
